Add a new folder modules with js files
