+++ src/modules/deleteCheck.js
@@ -1,0 +1,28 @@
+import {arrList, render, save} from '../index';
+
+export default function deleteCheck(ev) {
+    const elipseBtn = document.querySelectorAll('.elipse-btn');
+    const trashBtn = document.querySelectorAll('.trash-btn');
+    const item = ev.target;
+    if (item.classList[0] === 'check') {
+      const index = item.id;
+      const parent = item.parentElement;
+      parent.classList.toggle('line-through');
+  
+      elipseBtn[index].classList.toggle('hide');
+      trashBtn[index].classList.toggle('hide');
+      arrList[index].completed = !arrList[index].completed;
+      save();
+    }
+    if (item.classList[0] === 'trash-btn') {
+      const index = item.previousElementSibling.id;
+      const parent = item.parentElement;
+      parent.remove();
+      arrList.splice(index, 1);
+      arrList.forEach((object, index) => {
+        object.index = index;
+      });
+      save();
+      render();
+    }
+  }
